// src/generated/retail.ts
export interface paths {
    "/v1/customer/policy": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /**
         * Create customer enrollment
         * @description Provides approved partners the ability to call and send new Dental enrollments to the Guardian Retail administrative system. This will include necessary subscriber/member & dependent information, plan selection, etc.
         */
        post: {
            parameters: {
                query: {
                    /** @description Unique identifier provided by the client app each time they authorize and request for a new JWT bearer token. This is the same nonce that is used while requesting for JWT bearer token. */
                    nonce: components["schemas"]["Nonce"];
                };
                header: {
                    /** @description Unique id passed by the client to trace the request. Different for each and every request */
                    "x-correlation-id": components["schemas"]["CorrelationId"];
                    /** @description Unique id assigned as part of the partner onboarding process */
                    "x-partner-id": components["schemas"]["PartnerId"];
                    /** @description Source name assigned as a part of the partner onboarding process */
                    "x-source": components["schemas"]["Source"];
                };
                path?: never;
                cookie?: never;
            };
            /** @description Enroll customer request body */
            requestBody?: {
                content: {
                    "application/json": components["schemas"]["EnrollNewCustomerRequest"];
                };
            };
            responses: {
                /** @description Request received and is queued up for processing */
                202: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["AsyncResponse"];
                    };
                };
                /** @description Error happened while parsing the request */
                400: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
                /** @description Invalid request body */
                422: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["422ErrorResponse"];
                    };
                };
                /** @description Internal server error happened while processing the request */
                500: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["500ErrorResponse"];
                    };
                };
            };
        };
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/v1/customer/{primary-customer-number}/policy/{policy-number}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        /** Allows approved partners to send policy cancelation information for existing Dental policies. This information will update the corresponding policy in the Guardian Retail Admin System (this will cancel primary subscriber & all participants. If you want to cancel only a dependent, please use Terminate Dependent API) */
        put: {
            parameters: {
                query: {
                    /** @description Unique identifier provided by the client app each time they authorize and request for a new JWT bearer token. This is the same nonce that is used while requesting for JWT bearer token. */
                    nonce: components["schemas"]["Nonce"];
                };
                header: {
                    /** @description Unique correlation id for each request */
                    "x-correlation-id": components["schemas"]["CorrelationId"];
                    /** @description Unique id assigned as part of the partner onboarding process */
                    "x-partner-id": components["schemas"]["PartnerId"];
                    /** @description Source name assigned as a part of the partner onboarding process */
                    "x-source": components["schemas"]["Source"];
                };
                path: {
                    /** @description The primary customer number */
                    "primary-customer-number": components["schemas"]["CustomerNumber"];
                    /** @description The policy number for which the updates should be made to */
                    "policy-number": components["schemas"]["PolicyNumber"];
                };
                cookie?: never;
            };
            /** @description Update customer request */
            requestBody?: {
                content: {
                    "application/json": components["schemas"]["UpdateCustomerRequest"];
                };
            };
            responses: {
                /** @description Request received and is queued up for processing */
                202: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["AsyncResponse"];
                    };
                };
                /** @description Primary customer or associated policy not found */
                404: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
                /** @description Invalid request body */
                422: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["422ErrorResponse"];
                    };
                };
                /** @description Internal server error happened while processing the request */
                500: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["500ErrorResponse"];
                    };
                };
            };
        };
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/v1/customer/{primary-customer-number}/policy/{policy-number}/cancellation": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /**
         * Cancel an existing customer policy
         * @description The Retail Delete Dependent API allows approved partners the ability to delete dependents (spouse, children, etc.) from existing Dental policies. This information will update the corresponding policy in the Guardian Retail Admin system.
         */
        post: {
            parameters: {
                query: {
                    /** @description Unique identifier provided by the client app each time they authorize and request for a new JWT bearer token. This is the same nonce that is used while requesting for JWT bearer token */
                    nonce: components["schemas"]["Nonce"];
                };
                header: {
                    /** @description Unique correlation id for each request */
                    "x-correlation-id": components["schemas"]["CorrelationId"];
                    /** @description Unique id assigned as part of the partner onboarding process */
                    "x-partner-id": components["schemas"]["PartnerId"];
                    /** @description Source name assigned as a part of the partner onboarding process */
                    "x-source": components["schemas"]["Source"];
                };
                path: {
                    /** @description primary insured customer number the policy */
                    "primary-customer-number": components["schemas"]["CustomerNumber"];
                    /** @description The policy number for which the dependent has to be deleted from */
                    "policy-number": components["schemas"]["PolicyNumber"];
                };
                cookie?: never;
            };
            /** @description Cancel policy request body */
            requestBody?: {
                content: {
                    "application/json": components["schemas"]["CancelPolicyRequest"];
                };
            };
            responses: {
                /** @description Request received and is queued up for processing */
                202: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["AsyncResponse"];
                    };
                };
                /** @description Error happened while parsing the request */
                400: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
                /** @description Invalid request body */
                422: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["422ErrorResponse"];
                    };
                };
                /** @description Internal server error happened while processing the request */
                500: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["500ErrorResponse"];
                    };
                };
                /** @description Internal Error */
                default: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ErrorResponse"];
                    };
                };
            };
        };
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
// ...
        AsyncResponse: {
            correlationId?: components["schemas"]["CorrelationId"];
            primaryCustomerNumber?: components["schemas"]["CustomerNumber"];
            policyNumber?: components["schemas"]["PolicyNumber"];
            /** @example 2021-01-23T15:36:40.881 */
            requestAcceptedDateTime?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}